+++ components/camel-pgevent/src/test/java/org/apache/camel/pgevent/integration/PgEventPubSubIT.java
@@ -14,10 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.pgevent;
+package org.apache.camel.pgevent.integration;
 
-import com.impossibl.postgres.jdbc.PGDataSource;
-import org.apache.camel.BindToRegistry;
 import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
 import org.apache.camel.RoutesBuilder;
@@ -25,13 +23,7 @@ import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
-public class PgEventWithDefinedDatasourceTest extends PgEventTestSupport {
-
-    @EndpointInject("pgevent:///postgres/testchannel?datasource=#pgDataSource")
-    private Endpoint subscribeEndpoint;
-
-    @EndpointInject("pgevent:///postgres/testchannel?datasource=#pgDataSource")
-    private Endpoint notifyEndpoint;
+public class PgEventPubSubIT extends PgEventITSupport {
 
     @EndpointInject("timer://test?repeatCount=1&period=1")
     private Endpoint timerEndpoint;
@@ -39,20 +31,8 @@ public class PgEventWithDefinedDatasourceTest extends PgEventTestSupport {
     @EndpointInject("mock:result")
     private MockEndpoint mockEndpoint;
 
-    @BindToRegistry("pgDataSource")
-    public PGDataSource loadDataSource() throws Exception {
-        PGDataSource dataSource = new PGDataSource();
-        dataSource.setHost(getHost());
-        dataSource.setPort(getMappedPort());
-        dataSource.setDatabaseName(POSTGRES_DB);
-        dataSource.setUser(POSTGRES_USER);
-        dataSource.setPassword(POSTGRES_PASSWORD);
-
-        return dataSource;
-    }
-
     @Test
-    public void testPgEventPublishSubscribeWithDefinedDatasource() throws Exception {
+    public void testPgEventPublishSubscribe() throws Exception {
         mockEndpoint.expectedBodiesReceived(TEST_MESSAGE_BODY);
         mockEndpoint.assertIsSatisfied(5000);
     }
@@ -64,10 +44,13 @@ public class PgEventWithDefinedDatasourceTest extends PgEventTestSupport {
             public void configure() throws Exception {
                 from(timerEndpoint)
                         .setBody(constant(TEST_MESSAGE_BODY))
-                        .to(notifyEndpoint);
+                        .to(String.format(
+                                "pgevent://%s:%s/%s/testchannel?user=%s&pass=%s", getHost(), getMappedPort(), POSTGRES_DB,
+                                POSTGRES_USER, POSTGRES_PASSWORD));
 
-                from(subscribeEndpoint)
-                        .to(mockEndpoint);
+                from(String.format("pgevent://%s:%s/%s/testchannel?user=%s&pass=%s",
+                        getHost(), getMappedPort(), POSTGRES_DB, POSTGRES_USER, POSTGRES_PASSWORD))
+                                .to(mockEndpoint);
             }
         };
     }
